fix: updated the modal and modal screen export
