+++ packages/design-system/modal/modal.screen.tsx
@@ -1,3 +1,3 @@
-import { Modal } from "./index";
+import { Modal } from "./modal";
 
 export const ModalScreen = Modal;
